Refactor code

+++ Movie.API/Infrastructure/Repositories/RoleRepository.cs
@@ -10,29 +10,30 @@ namespace Movie.API.Infrastructure.Repositories
     {
         Task<bool> DeleteRoleAsync(string rolename);
     }
-   
+
     public class RoleRepository : GenericRepository<Role>, IRoleRepository
     {
         private readonly MovieDbContext _dbContext;
         public readonly RoleManager<Role> _roleManager;
-        public RoleRepository(MovieDbContext context , RoleManager<Role> roleManager) : base(context)
+        public RoleRepository(MovieDbContext context, RoleManager<Role> roleManager) : base(context)
         {
             _dbContext = context;
             _roleManager = roleManager;
         }
-        public Task<Role> AddAsync(Role entity)
+        public async new Task<Role> AddAsync(Role entity)
         {
-            _roleManager.CreateAsync(entity).Wait();
-            return Task.FromResult(entity);
+            await _roleManager.CreateAsync(entity);
+            return entity;
         }
-        public Task<Role> UpdateAsync(Role entity)
+        public async new Task<Role> UpdateAsync(Role entity)
         {
-            _roleManager.UpdateAsync(entity).Wait();
-            return Task.FromResult(entity);
+            await _roleManager.UpdateAsync(entity);
+            return entity;
         }
         public async Task<bool> DeleteRoleAsync(string rolename)
         {
             var role = await _roleManager.FindByNameAsync(rolename);
+            if (role is null) return false;
             await _roleManager.DeleteAsync(role);
             return true;
         }
